+++ app/src/main/java/protect/card_locker/BarcodeValuesListDisambiguatorCallback.java
@@ -1,0 +1,6 @@
+package protect.card_locker;
+
+public interface BarcodeValuesListDisambiguatorCallback {
+    void onUserChoseBarcode(BarcodeValues barcodeValues);
+    void onUserDismissedSelector();
+}