+++ app/src/main/java/protect/card_locker/Utils.java
@@ -12,8 +12,10 @@ import android.graphics.BitmapFactory;
 import android.graphics.Color;
 import android.graphics.ImageDecoder;
 import android.graphics.Matrix;
+import android.graphics.pdf.PdfRenderer;
 import android.net.Uri;
 import android.os.Build;
+import android.os.ParcelFileDescriptor;
 import android.provider.MediaStore;
 import android.text.Layout;
 import android.text.Spanned;
@@ -39,6 +41,7 @@ import androidx.exifinterface.media.ExifInterface;
 import androidx.palette.graphics.Palette;
 
 import com.google.android.material.color.DynamicColors;
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.google.zxing.BinaryBitmap;
 import com.google.zxing.LuminanceSource;
 import com.google.zxing.MultiFormatReader;
@@ -46,6 +49,8 @@ import com.google.zxing.NotFoundException;
 import com.google.zxing.RGBLuminanceSource;
 import com.google.zxing.Result;
 import com.google.zxing.common.HybridBinarizer;
+import com.google.zxing.multi.GenericMultipleBarcodeReader;
+import com.google.zxing.multi.MultipleBarcodeReader;
 
 import java.io.BufferedReader;
 import java.io.ByteArrayOutputStream;
@@ -64,6 +69,7 @@ import java.text.NumberFormat;
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Collections;
 import java.util.Currency;
 import java.util.Date;
 import java.util.GregorianCalendar;
@@ -83,12 +89,13 @@ public class Utils {
     public static final int SELECT_BARCODE_REQUEST = 2;
     public static final int BARCODE_SCAN = 3;
     public static final int BARCODE_IMPORT_FROM_IMAGE_FILE = 4;
-    public static final int CARD_IMAGE_FROM_CAMERA_FRONT = 5;
-    public static final int CARD_IMAGE_FROM_CAMERA_BACK = 6;
-    public static final int CARD_IMAGE_FROM_CAMERA_ICON = 7;
-    public static final int CARD_IMAGE_FROM_FILE_FRONT = 8;
-    public static final int CARD_IMAGE_FROM_FILE_BACK = 9;
-    public static final int CARD_IMAGE_FROM_FILE_ICON = 10;
+    public static final int BARCODE_IMPORT_FROM_PDF_FILE = 5;
+    public static final int CARD_IMAGE_FROM_CAMERA_FRONT = 6;
+    public static final int CARD_IMAGE_FROM_CAMERA_BACK = 7;
+    public static final int CARD_IMAGE_FROM_CAMERA_ICON = 8;
+    public static final int CARD_IMAGE_FROM_FILE_FRONT = 9;
+    public static final int CARD_IMAGE_FROM_FILE_BACK = 10;
+    public static final int CARD_IMAGE_FROM_FILE_ICON = 11;
 
     public static final String CARD_IMAGE_FILENAME_REGEX = "^(card_)(\\d+)(_(?:front|back|icon)\\.png)$";
 
@@ -131,6 +138,80 @@ public class Utils {
         return ColorUtils.calculateLuminance(backgroundColor) > LUMINANCE_MIDPOINT;
     }
 
+    static public List<BarcodeValues> retrieveBarcodesFromImage(Context context, Uri uri) {
+        Log.i(TAG, "Received image file with possible barcode");
+
+        if (uri == null) {
+            Log.e(TAG, "Uri did not contain any data");
+            Toast.makeText(context, R.string.errorReadingImage, Toast.LENGTH_LONG).show();
+            return new ArrayList<>();
+        }
+
+        Bitmap bitmap;
+        try {
+            bitmap = retrieveImageFromUri(context, uri);
+        } catch (IOException e) {
+            Log.e(TAG, "Error getting data from image file");
+            e.printStackTrace();
+            Toast.makeText(context, R.string.errorReadingImage, Toast.LENGTH_LONG).show();
+            return new ArrayList<>();
+        }
+
+        List<BarcodeValues> barcodesFromBitmap = getBarcodesFromBitmap(bitmap);
+
+        if (barcodesFromBitmap.isEmpty()) {
+            Log.i(TAG, "No barcode found in image file");
+            Toast.makeText(context, R.string.noBarcodeFound, Toast.LENGTH_LONG).show();
+        }
+
+        return barcodesFromBitmap;
+    }
+
+    static public List<BarcodeValues> retrieveBarcodesFromPdf(Context context, Uri uri) {
+        Log.i(TAG, "Received PDF file with possible barcode");
+
+        if (uri == null) {
+            Log.e(TAG, "Uri did not contain any data");
+            Toast.makeText(context, R.string.errorReadingFile, Toast.LENGTH_LONG).show();
+            return new ArrayList<>();
+        }
+
+        ParcelFileDescriptor parcelFileDescriptor;
+        PdfRenderer renderer;
+        try {
+            parcelFileDescriptor = context.getContentResolver().openFileDescriptor(uri, "r");
+            renderer = new PdfRenderer(parcelFileDescriptor);
+        } catch (IOException e) {
+            Log.e(TAG, "Could not read file in uri");
+            Toast.makeText(context, R.string.errorReadingFile, Toast.LENGTH_LONG).show();
+            return new ArrayList<>();
+        }
+
+        // Loop over all pages to find barcodes
+        List<BarcodeValues> barcodesFromPdfPages = new ArrayList<>();
+        Bitmap renderedPage;
+        for (int i = 0; i < renderer.getPageCount(); i++) {
+            PdfRenderer.Page page = renderer.openPage(i);
+            renderedPage = Bitmap.createBitmap(page.getWidth(), page.getHeight(), Bitmap.Config.ARGB_8888);
+            page.render(renderedPage, null, null, PdfRenderer.Page.RENDER_MODE_FOR_DISPLAY);
+            page.close();
+
+            List<BarcodeValues> barcodesFromPage = getBarcodesFromBitmap(renderedPage);
+            for (BarcodeValues barcodeValues : barcodesFromPage) {
+                barcodeValues.setNote(String.format(context.getString(R.string.pageWithNumber), i+1));
+                barcodesFromPdfPages.add(barcodeValues);
+            }
+        }
+        renderer.close();
+
+        if (barcodesFromPdfPages.isEmpty()) {
+            Log.i(TAG, "No barcode found in pdf file");
+            Toast.makeText(context, R.string.noBarcodeFound, Toast.LENGTH_LONG).show();
+        }
+
+        return barcodesFromPdfPages;
+    }
+
     /**
      * Returns the Barcode format and content based on the result of an activity.
      * It shows toasts to notify the end-user as needed itself and will return an empty
@@ -142,45 +223,20 @@ public class Utils {
      * @param context
      * @return BarcodeValues
      */
-    static public BarcodeValues parseSetBarcodeActivityResult(int requestCode, int resultCode, Intent intent, Context context) {
+    static public List<BarcodeValues> parseSetBarcodeActivityResult(int requestCode, int resultCode, Intent intent, Context context) {
         String contents;
         String format;
 
         if (resultCode != Activity.RESULT_OK) {
-            return new BarcodeValues(null, null);
+            return new ArrayList<>();
         }
 
         if (requestCode == Utils.BARCODE_IMPORT_FROM_IMAGE_FILE) {
-            Log.i(TAG, "Received image file with possible barcode");
+            return retrieveBarcodesFromImage(context, intent.getData());
+        }
 
-            Uri data = intent.getData();
-            if (data == null) {
-                Log.e(TAG, "Intent did not contain any data");
-                Toast.makeText(context, R.string.errorReadingImage, Toast.LENGTH_LONG).show();
-                return new BarcodeValues(null, null);
-            }
-
-            Bitmap bitmap;
-            try {
-                bitmap = retrieveImageFromUri(context, data);
-            } catch (IOException e) {
-                Log.e(TAG, "Error getting data from image file");
-                e.printStackTrace();
-                Toast.makeText(context, R.string.errorReadingImage, Toast.LENGTH_LONG).show();
-                return new BarcodeValues(null, null);
-            }
-
-            BarcodeValues barcodeFromBitmap = getBarcodeFromBitmap(bitmap);
-
-            if (barcodeFromBitmap.isEmpty()) {
-                Log.i(TAG, "No barcode found in image file");
-                Toast.makeText(context, R.string.noBarcodeFound, Toast.LENGTH_LONG).show();
-            }
-
-            Log.i(TAG, "Read barcode id: " + barcodeFromBitmap.content());
-            Log.i(TAG, "Read format: " + barcodeFromBitmap.format());
-
-            return barcodeFromBitmap;
+        if (requestCode == Utils.BARCODE_IMPORT_FROM_PDF_FILE) {
+            return retrieveBarcodesFromPdf(context, intent.getData());
         }
 
         if (requestCode == Utils.BARCODE_SCAN || requestCode == Utils.SELECT_BARCODE_REQUEST) {
@@ -196,7 +252,7 @@ public class Utils {
             Log.i(TAG, "Read barcode id: " + contents);
             Log.i(TAG, "Read format: " + format);
 
-            return new BarcodeValues(format, contents);
+            return Collections.singletonList(new BarcodeValues(format, contents));
         }
 
         throw new UnsupportedOperationException("Unknown request code for parseSetBarcodeActivityResult");
@@ -216,22 +272,22 @@ public class Utils {
         return MediaStore.Images.Media.getBitmap(context.getContentResolver(), data);
     }
 
-    static public BarcodeValues getBarcodeFromBitmap(Bitmap bitmap) {
+    static public List<BarcodeValues> getBarcodesFromBitmap(Bitmap bitmap) {
         // This function is vulnerable to OOM, so we try again with a smaller bitmap is we get OOM
         for (int i = 0; i < 10; i++) {
             try {
-                return Utils.getBarcodeFromBitmapReal(bitmap);
+                return Utils.getBarcodesFromBitmapReal(bitmap);
             } catch (OutOfMemoryError e) {
-                Log.w(TAG, "Ran OOM in getBarcodeFromBitmap! Trying again with smaller picture! Retry " + i + " of 10.");
+                Log.w(TAG, "Ran OOM in getBarcodesFromBitmap! Trying again with smaller picture! Retry " + i + " of 10.");
                 bitmap = Bitmap.createScaledBitmap(bitmap, (int) Math.round(0.75 * bitmap.getWidth()), (int) Math.round(0.75 * bitmap.getHeight()), false);
             }
         }
 
         // Give up
-        return new BarcodeValues(null, null);
+        return new ArrayList<>();
     }
 
-    static private BarcodeValues getBarcodeFromBitmapReal(Bitmap bitmap) {
+    static private List<BarcodeValues> getBarcodesFromBitmapReal(Bitmap bitmap) {
         // In order to decode it, the Bitmap must first be converted into a pixel array...
         int[] intArray = new int[bitmap.getWidth() * bitmap.getHeight()];
         bitmap.getPixels(intArray, 0, bitmap.getWidth(), 0, 0, bitmap.getWidth(), bitmap.getHeight());
@@ -240,13 +296,61 @@ public class Utils {
         LuminanceSource source = new RGBLuminanceSource(bitmap.getWidth(), bitmap.getHeight(), intArray);
         BinaryBitmap binaryBitmap = new BinaryBitmap(new HybridBinarizer(source));
 
+        List<BarcodeValues> barcodeValuesList = new ArrayList<>();
         try {
-            Result barcodeResult = new MultiFormatReader().decode(binaryBitmap);
+            MultiFormatReader multiFormatReader = new MultiFormatReader();
+            MultipleBarcodeReader multipleBarcodeReader = new GenericMultipleBarcodeReader(multiFormatReader);
 
-            return new BarcodeValues(barcodeResult.getBarcodeFormat().name(), barcodeResult.getText());
+            Result[] barcodeResults = multipleBarcodeReader.decodeMultiple(binaryBitmap);
+
+            for (Result barcodeResult : barcodeResults) {
+                Log.i(TAG, "Read barcode id: " + barcodeResult.getText());
+                Log.i(TAG, "Read format: " + barcodeResult.getBarcodeFormat().name());
+
+                barcodeValuesList.add(new BarcodeValues(barcodeResult.getBarcodeFormat().name(), barcodeResult.getText()));
+            }
+
+            return barcodeValuesList;
         } catch (NotFoundException e) {
-            return new BarcodeValues(null, null);
+            return barcodeValuesList;
         }
+    }
+
+    static public void makeUserChooseBarcodeFromList(Context context, List<BarcodeValues> barcodeValuesList, BarcodeValuesListDisambiguatorCallback callback) {
+        // If there is only one choice, consider it chosen
+        if (barcodeValuesList.size() == 1) {
+            callback.onUserChoseBarcode(barcodeValuesList.get(0));
+            return;
+        }
+
+        // Ask user to choose a barcode
+        // TODO: This should contain an image of the barcode in question to help users understand the choice they're making
+        CharSequence[] barcodeDescriptions = new CharSequence[barcodeValuesList.size()];
+        for (int i = 0; i < barcodeValuesList.size(); i++) {
+            BarcodeValues barcodeValues = barcodeValuesList.get(i);
+            CatimaBarcode catimaBarcode = CatimaBarcode.fromName(barcodeValues.format());
+
+            String barcodeContent = barcodeValues.content();
+            // Shorten overly long barcodes
+            if (barcodeContent.length() > 22) {
+                barcodeContent = barcodeContent.substring(0, 20) + "…";
+            }
+
+            if (barcodeValues.note() != null) {
+                barcodeDescriptions[i] = String.format("%s: %s (%s)", barcodeValues.note(), catimaBarcode.prettyName(), barcodeContent);
+            } else {
+                barcodeDescriptions[i] = String.format("%s (%s)", catimaBarcode.prettyName(), barcodeContent);
+            }
+        }
+
+        MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(context);
+        builder.setTitle(context.getString(R.string.multipleBarcodesFoundPleaseChooseOne));
+        builder.setItems(
+                barcodeDescriptions,
+                (dialogInterface, i) -> callback.onUserChoseBarcode(barcodeValuesList.get(i))
+        );
+        builder.setOnCancelListener(dialogInterface -> callback.onUserDismissedSelector());
+        builder.show();
     }
 
     static public Boolean isNotYetValid(Date validFromDate) {